There's bug somewhere
The set of blip with address lookup works only at the start of work on
the page or after you clear address in left column, not in central one.

In central column blip set up works **after you clear address AND press
ENTER** so that field is marked as not entered in the list at right.

I don't know what is going on.